⚗️ Attempt some conversion adjustments found in the INA237 spec sheet
**NOTE**: This change also ran `clang-format` on the sources so they
follow the upstream esphome formatting guidelines.

These conversions were initially missed, and because we're several
layers of abstraction away, this is intended to serve as more of a
playground so we can get at least *one* value correct, specifically that
of the shunt voltage. Once done, we'll be able to apply these same
conversions elsewhere.

**NOTE**: It is currently unknown if ESPHome is properly treating our
values as big or little endian, so I don't expect values to be correct
most of the time by default.